+++ examples/server.php
@@ -3,7 +3,7 @@
  * Non-Blocking Server
  */
 prggmr\load_module('socket');
-prggmr\load_module('time');
+var_dump(prggmr\load_module('time'));
 
 $server = new prggmr\module\socket\server\Select("0.0.0.0:1337");
 
@@ -13,16 +13,14 @@ $server->on_connect(new prggmr\Handle(function(){
     if (!isset($this->server->count)) {
         $this->server->count = 0;
     } else {
-        echo $this->server->count.PHP_EOL;
         $this->server->count++;
-        if ($this->server->count >= 500) {
+        if ($this->server->count >= 100) {
             // If more than 500 after this
             prggmr\after($this->get_signal(), function() use ($server){
                 $server->reconnect();
                 $server->count = 0;
             });
         }
-        echo $this->server->count.PHP_EOL;
     }
     $this->write("You sent me the following".PHP_EOL);
     $this->write($this->read());
@@ -30,7 +28,7 @@ $server->on_connect(new prggmr\Handle(function(){
 
 // // On Disconnect
 $server->on_disconnect(new prggmr\Handle(function(){
-    echo "Disconnecting".PHP_EOL;
+    // echo "Disconnecting".PHP_EOL;
 }, null));
 
 // Register the server
@@ -38,6 +36,6 @@ prggmr\handle($server, function(){
     echo "Server is running at ".$this->get_address().PHP_EOL;
 });
 
-prggmr\module\time\interval(10, function(){
-    echo "10 seconds goes by...".PHP_EOL;
+prggmr\module\time\interval(1, function() use ($server){
+    echo $server->count.PHP_EOL;
 }, prggmr\engine\idle\Time::SECONDS);