+++ src/engine/templates/html.php
@@ -1,9 +1,0 @@
-<!doctype html>
-<html>
-<head>
-    <title>prggmr Event Analysis</title>
-</head>
-<body>
-<h1>prggmr Event Analysis</h1>
-<hr />
-</body>